#4 - hotfixes method sumOfDigits()

+++ input.java
@@ -5,12 +5,12 @@ class Main {
         Scanner input = new Scanner(System.in);
         int n = input.nextInt();
         System.out.println(n);
-        int sum = sumOfDigits(n);
+        long sum = sumOfDigits(n);
         System.out.print(sum);
     }
 
-    public static int sumOfDigits(int n) {
-        int sum = 0;
+    public static long sumOfDigits(int n) {
+        long sum = 0;
         do {
             sum += n % 10;
             n /= 10;
